+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to Toisto will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Explicitly distinguish between grammatical aspect (perfective and imperfective) and grammatical tense (present tense and past tense). Closes [#1147](https://github.com/fniessink/toisto/issues/1147).
+
 ## 0.38.2 - 2025-09-19
 
 ### Fixed